app/src/main/python/requirements.txt
pytesseract
opencv-python-headless
Pillow